requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.1.7
django-allauth==0.44.0
django-bootstrap4==2.3.1
django-heroku==0.3.1
django-import-export==2.5.0
et-xmlfile==1.0.1
gunicorn==20.0.4
idna==2.10
MarkupPy==1.14
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.7
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.0.1
python-dotenv==0.15.0
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.2
sqlparse==0.4.1
tablib==3.0.0
urllib3==1.26.3
whitenoise==5.2.0
xlrd==2.0.1
xlwt==1.3.0
